backtesting/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28

# Market data providers
alpaca-py>=0.9.0

# Technical indicators
ta-lib>=0.4.28
pandas-ta>=0.3.14b0

# Performance metrics
quantstats>=0.0.62
empyrical>=0.5.5

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0
